% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotka_volterra_preypredator.R
\name{lotka_volterra_preypredator}
\alias{lotka_volterra_preypredator}
\title{Lotka-Volterra Predator-Prey Model}
\usage{
lotka_volterra_preypredator(initial_conditions, parameters, time)
}
\arguments{
\item{initial_conditions}{A vector of initial conditions for the model stating the initial population of prey (x) and predator (y).}

\item{parameters}{A vector of parameters for the model stating the prey growth rate (a), predator death rate (b), predator's efficiency at converting food to new predators (c), and the predator's rate of consumption of prey (d).}

\item{time}{A vector of time points to solve the model at}
}
\value{
A data frame with the time points and the population of prey and predator at each time point.
}
\description{
Lotka-Volterra Predator-Prey Model
}
\examples{
initial_conditions <- c(x = 40, y = 9)
parameters <- c(a = 0.1, b = 0.02, c = 0.0005, d = 0.3)
time <- seq(0, 200, by = 0.1)
results_df <- lotka_volterra_preypredator(initial_conditions, parameters, time)
}
